ELB_DDoS_Defender/requirements.txt
boto3>=1.26.0
scapy>=2.5.0
pyyaml>=6.0
jinja2>=3.1.0
weasyprint>=59.0
reportlab>=4.0.0
requests>=2.28.0
rich>=13.0.0
pyshark>=0.6
